+++ Libraries/TGaus.h
@@ -10,8 +10,8 @@ class TGaus
 {
 private:
 
-	std::vector<std::vector<float>>				m_system;
-	std::vector<float>							m_solution;
+	std::vector<std::vector<double>>				m_system;
+	std::vector<double>							m_solution;
 
 public:
 
@@ -21,9 +21,9 @@ public:
 
 	void										createSystemDrivers( const std::vector<TLink>& drivers );
 
-	std::vector<std::vector<float>>				getSystem() const;
+	std::vector<std::vector<double>>				getSystem() const;
 
-	const std::vector<float>&					getSolution() const;
+	const std::vector<double>&					getSolution() const;
 
 };
